+++ app/Http/Controllers/Admin/Recipe/RecipeTraitDelete.php
@@ -5,17 +5,20 @@ use Illuminate\Http\Request;
 
 trait RecipeTraitDelete
 {
-    public function delete(Request $request, $recipe_id)
+    public function delete(Request $request, string $recipe_id) : ?\Illuminate\Http\RedirectResponse
     {
         $row = \App\Models\Recipe::find($recipe_id);
 
-        // ユーザ削除はactiveをoffに。
-        if(!$row->delete()) {
-            // 保存失敗
-            \U::invokeErrorValidate($request, "削除に失敗しました。");
+        if($row) {
+            // ユーザ削除はactiveをoffに。
+            if(!$row->delete()) {
+                // 保存失敗
+                \U::invokeErrorValidate($request, "削除に失敗しました。");
+            }
+            // それぞれのroleに応じた一覧へ移動。
+            return redirect()->route("admin-recipe-index")->with("message-success", "削除しました。");
         }
-        // それぞれのroleに応じた一覧へ移動。
-        return redirect()->route("admin-recipe-index")->with("message-success", "削除しました。");
+        return null;
     }
 
     // *************************************
